move all code to plugin/virtualenv.vim
This change allows the user to immediately activate a virtual
environment without needing to edit a file whatsoever. For example, the
following is now possible:

$ vim
:VirtualEnvActivate myenvironment

Previously the only time vim-virtualenv would be loaded is if a Python
file was opened. However, there are use cases where you want to load a
virtualenv into VIM's internal Python interpreter even when not working
on Python files at all. For example, a vim plugin may require a special
module that the user has chosen to install into a virtualenv. In this
case, the user needs to activate the virtualenv in order for the vim
plugin to work correctly while editing other types of files (rst for
example).