Add relative path support for `g:virtualenv_directory`
If a relative path is set, the plugin will check that this relative path
exists first before running any of it's functions. If the pat does not
exist, then it will gracefully default back to the value of
`$WORKON_HOME` first, then `~/.virtualenvs` last. This is is done based
on what vim's CWD is.