Add check for Pipenv when auto-detecting virtualenv
If the executable Pipenv exists, and returns a path when called, use the
output as the path to the virtual environment.

The order of the checks could be challenged, here, but personally I
believe that if Pipenv has an environment for the current directory, it
should probably be the active env. This makes it so you don't have to
worry about what else is in your environment when you change into the
project directory.